chore: Compute function tree root in ts. (#3326)
One step closer to not needing circuits.wasm.
* This creates a new class `MerkleTreeRootCalculator`. It wouldn't have
had to exist except for the fact that the zero leaf of the function tree
is the hash of a zeroed function leaf (5 fields) rather than just a 0
buffer. Kinda annoying. Needed?
* I tried putting the class in the merkle-tree package, but worryingly
that created a circular dependency because everything depends on this
generic "types" catch all package, and it depends on circuits... Seems
sus. For now the class sits next to where it's used in `abis`.
* When performing a naive tree hash, this actually beats the wasm
performance by around 5%. Perhaps due to the wasm always computing the
zero leaf. Assuming that's the reason, the performance is equal despite
overhead of calling into wasm (so we can assume such overheads are
negligible).
* The algorithm performs much better than the wasm version however, when
the tree is not full, as it leverages zero layer caches to not have to
hash the entire tree. The test adds 4 leaves to a 16 leaf tree, and as
expected is 4 times faster.

Just for clarity, this still uses wasm, but it uses bb.js wasm and at a
lower level. Just calling `pedersenHash`.