chore: all hashes in ts (#3333)
This is mostly just removing a bunch of async/await stuff as we convert
the final 3 functions in abis.ts to use `foundation/crypto/pedersen`.

There are two things that are iffy, but not sure if that should block a
merge as everything passes:
* `vkHash`, I've hardcoded a couple of values I don't fully understand
(coset generator and root, which are pulled from evalutation domain in
the C++).
* `computePublicCallStackItemHash` the C++ did something funky when
`isExecutionRequest` is true. I don't fully understand what and if it
needs reproducing here, but again, everything passes.

Have left the commented C++ code in the two places above for reference.