feat: fold batching challenge (alpha) (#3291)
This PR provides an initial version of folding the subrelation batching
challenge `alpha`. The future plan is to have one `alpha` per
subrelation to ensure the degree of the combiner stays reasonable, to be
handled in a future PR . and so, for now, we use one `alpha` per
Instance, fold them and use the same folded alpha to batch each
subrelation in the combiner.

I separated some code between sumcheck and protogalaxy because
supporting both was starting to make it too complicated. Also `alpha` is
now passed to the sumcheck protocol rather than produced inside the
protocol to facilitate extracting and including as part of the
instances.