chore(ci): Fix publishing of ACVM crates (#3645)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR does an "I reckon" fix to publishing ACVM crates.

I've added a version number to the workspace `Cargo.toml` so that these
crates can be pushed to crates.io. The issue then is that we require
these to be bumped appropriately by the release-please PR.

I imagine the reason why this wasn't done originally is that it requires
the `acvm-repo` release-please package to change files outside of its
directory. I've given this a go however and we can check whether the PR
gets updated correctly after this is merged.

cc @kevaundray as he set up the original release please setup for ACVM.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.