chore: adding docs to release please (#3571)
# Description

This PR attempts to add docs to release_please

## Problem\*

So far our attempts with docs workflows have been unsuccessful. The
latest try was almost there, but it implied having a PR after the
release is merged, which would actually make the docs show up _after_
the tag, not before. So when the release was tagged as stable, it
wouldn't have the docs at that point in time.

## Summary\*

This change attempts to do some blind changes, as I tried faking master
with another branch and somehow didn't make it actually trigger new
tags.

---------

Co-authored-by: Tom French <tom@tomfren.ch>