feat: data bus (#3508)
# Description

## Problem\*

aztec private kernel is doing a lot of proof recursion along the call
stack of contract calls, resulting in a lot of public inputs required
for the aztec ABI. These amount of public inputs have an impact on the
performance, which can be mitigated by having instead the public inputs
as private and a public commitment of these inputs. We can then leverage
the proving system capability for proving/verifying efficiently
commitments, outside of the noir program.

## Summary\*
In order to interface with the proving system, the user can declare some
inputs as call_data (instead of pub) and the return values as
return_data (instead of pub). The compiler will abstract these into
specific immutable arrays that the proving system will use.



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [X] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>