feat: throw compile time error if contract has too many fns (#3536)
fix #3500 

Contracts can be compiled using nargo and nargo-wasm. So either we place
compile time checks in aztec-nr or in `noir-compiler`. I have done it in
later - specifically when we generate contract artifacts since that is
the earliest function used by both noir-wasm + compiler.

Let me know if there is a better place to add the check 🤷