docs: documenting issue with `context.block_header` (#3565)
I was incorrectly trying to fetch a sibling path from an older block
when proving that a block header is legit. This caused the test to break
when proving inclusion at an older block. I currently could not address
it because of issue #3564 so I decided to only document it in this PR.