fix: Make trait functions generic over `Self` (#3702)
# Description

## Problem & Summary

Checking a trait function's type previously was very ad-hoc. When a
`TraitMethodReference` was found, it'd manually create a function type
from the trait function's definition and instantiate it - even though
instantiating a `Type::Function` does nothing.

I've changed the TraitFunction struct to add a `Type` field for the
whole function type, which includes generics. The type now should
usually be a `Type::Forall`. In addition, I've added the trait's `Self`
type to the list of generics in the Type::Forall.

This fixes the issue where the `Self` type would be bound after certain
function calls which would affect callsites but not a function body
leading to a type mismatch between function return type and a call's
return type. This is the issue we were encountering after adding traits
to noir-protocol-circuits.

## Additional Information

No tests are provided since I was unable to make a minimal repro from
the thousands of lines of noir-protocol-circuits unfortunately. I'll
keep experimenting while this PR is up.

I'm also merging into jf/fix-3089 since this was somewhat accidentally
built on top of it. The two changes are otherwise unrelated.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.