feat: docs landing page with a playground (#3667)
# Description

This PR adds a landing page for the docs with a playground

## Problem\*

Our first page was OK but could be improved (even had references to a
"book", as the previous version was a rust book haha). One idea was to
have a nice first page with an editable playground, so I hacked one in
Istanbul and refined it a bit over the last few days.

Plans are to update the whole design sometime soon, so I didn't put a
lot of care into its design.

## Summary\*

-  Adds an `index.mdx`
- Configures docusaurus to resolve webpack stuff
- Imports the playground npm package
- Moves some stuff out of docs/index page and formats it with prettier

## Additional Context

Playground is lazy-loaded and triggers only when user clicks a CTA
button. The reason is to avoid downloading quite some hefty wasm, which
if done synchronously would add a lot of build time