chore(ci): fail `tests-end` job if any dependants failed (#3737)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

It seems like despite the `tests-end` job not running and being skipped
if any of the dependants fail, this is enough to satisfy the required
check status. This PR changes it so that this job explicitly fails if
any of the dependants failed or were skipped.


## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

Co-authored-by: kevaundray <kevtheappdev@gmail.com>