feat: optimize out unnecessary truncation instructions (#3717)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We're currently adding a double truncation when performing overflow
checks on left shifts. We could/should address this directly but as a
more general rule we can optimize out any truncation which is truncating
a value to fit into a type which is equal or larger to the source type.

I've also replaced a few instances of `insert_instruction` with the more
instruction-specific version to reduce verbosity.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: kevaundray <kevtheappdev@gmail.com>