chore: re-export the items needed for the lsp from the fm crate instead of importing codespan_reporting (#3757)
# Description

This is a small change to control the scope of what the lsp touches and
to see what the fm crate is offering to other crates in this workspace.

codespan_reporting is needed because the fm crate is using codespan

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.