fix: Stop cloning Traits! (#3736)
# Description

## Problem\*

The `Trait` struct is quite large and was being cloned every time
`get_trait` was called in the NodeResolver, which was somewhat often.

## Summary\*

Stops the cloning of traits! The `Clone` derive is removed entirely and
when needed for ownership reasons we either re-retrieve the trait by
calling `get_trait` again (which now returns a reference), or we
temporarily take ownership of the trait's methods.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: kevaundray <kevtheappdev@gmail.com>