feat: reorganizing docs to fit diataxis framework (#3711)
# Description

One of DevRel goals was to slowly start designing our docs process, and
specially adopting the Diátaxis framework. A quick picture should make
it easy to understand what exactly is Diátaxias:

![diataxis](https://diataxis.fr/_images/diataxis.png)

Relevant WIP docs for this are (internal links):
- [docs philosophy](https://hackmd.io/0ffxsnAARZaaiygZD_kWEA)
- [Noir docs 2.0](https://hackmd.io/IdN1KMjIR3CgKk4t0A6-ug)

This PR is a first step towards that reorg, by essentially moving things
around. There are little changes to the content itself for now.

However, during the reorg process, an opportunity came to upgrade to
Docusaurus v3.0 with very little overhead, so I took it.

## Problem\*

A small retro was made as part of the above [Noir Docs
2.0](https://hackmd.io/IdN1KMjIR3CgKk4t0A6-ug) exercise.

## Summary\*

The Diátaxis framework divides docs neatly into four sections. However,
in this case, it was deemed important to give a quick reference, so as
to tackle the initial interest and motivation of readers and tinkerers.

So, five top-level sections:
- Getting Started -> Has a quick reference for all of the other
sections, by providing a guide, a tutorial, an explanation, and a small
reference to other tools. The idea is that most user journeys will
eventually start here.
- "The Noir Language" -> One of the biggest tasks will be to neatly
separate explanations from reference material. This is a task that will
be done later on. So this section contains a hybrid.
- How-To Guides -> Given the clear distinction between "How-To" and
"Tutorial", which is important to keep, this section currently doesn't
have any more than the potentially outdated Merkle Tree example.
- Tutorials -> Again, since no content was added, only the NoirJS e2e
example is in this section.
- Reference -> The reference material for Nargo commands and NoirJS
lives here. This will be gradually expanded to fit the reference in the
Language section.