chore: remove aztec compile time feature flag (#3596)
# Description

This is a followup to https://github.com/noir-lang/noir/pull/3578 where
we move the feature flag from compile time to runtime producing a single
binary.

EDIT: This was initially was meant to change it to runtime flag, but
there was quite a lot of effort to make this work. We can introduce an
environment variable in the future to turn off all macros if this turn
out to be a problem

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: Tom French <tom@tomfren.ch>