fix: remove `noirc_driver/aztec` feature flag in docker (#3784)
# Description

This was not caught because the docker tests are not required checks and
it was only being ran on master.

They are not required because they need to be optimized a bit more with
caching. I don't expect them to fail often, so in the interim, its okay
to not have these as required (If this happens again and we have not
migrated, then we can revisit)

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.