chore: moving ordering to category jsons and frontmatters (#3777)
# Description

This big PR does a simple thing: It makes pages ordering respect
metadata instead of filenames

## Problem\*

A naive approach supported over the years at docusaurus was to simply
have files prefixed with 00, 01, 02, which were removed automatically by
docusaurus.

This makes docs a bit harder to maintain if for some reason you decide
to move pages around, as it breaks all the links.

## Summary\*

- Renames all the folders and pages to remove the prefixes
- Adds `sidebar_position` property in frontmatter, to order them 
- Adds `_category_.json` files to order folders
- Removes ordering of top level folders in `sidebars.json`