fix: Stop issuing unused variable warnings for variables in trait definitions (#3797)
# Description

## Problem\*

Previously we'd issue a warning for each trait function parameter that
it is unused. This seemed to only trigger when the trait was in another
crate.

## Summary\*

Stop checking for unused variables in trait function declarations since
there is no body to use the variable in anyway.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

Co-authored-by: kevaundray <kevtheappdev@gmail.com>