chore: move module related code from `fm` to `noirc-frontend` (#3806)
# Description

## Problem\*

Resolves #3799 

## Summary\*

This is a relatively quick and dirty resolution to #3799. We may want to
integrate these functions into `ModCollector` as methods but I think
they also work standalone plus it means that we can transfer the tests
across without needing to create a `ModCollector` instance.

## Additional Context


## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.