chore: filtering patches from docs deployment script (#3779)
# Description

This change makes the script that runs before `yarn build` filter down
the patches to the latest patch of every major.

## Problem\*

This script used to remove:
- Prereleases
- Nightlies (which are prereleases)
- "Aztec" versions

However, this would mean that every patch, as long as it was stable,
would have its own docs deployed. This would lead to docs having three
patch versions, which shouldn't even differ (they're patches, no
breaking changes), like 0.19.1, 0.19.2, etc

## Summary\*

This change simplifies the script by removing some computation that was
there just for debugging reasons. It also stops assuming all "aztec"
versions are prereleases. And finally, it filters down only for the
latest patch version.


## Additional Context

Another way of doing it (which is how we're doing it now) is to just
make all non-latest patch as prereleases, which is kinda annoying. With
this change, it shouldn't matter.