fix: map relative path to protocol circuits (#3694)
This PR fixes an issue with how we mirror code in the monorepo to
aztecprotocol/aztec-nr.

A while back the root aztec library added a dependency on
noir-protocol-circuits. In the monorepo this is achieved by use of a
relative path in Nargo.toml. This works fine for contracts that pull in
Aztec.nr through the monorepo (see also #3604) but it breaks projects
that use
[AztecProtocol/aztec-nr](https://github.com/AztecProtocol/aztec-nr)
because the relative path won't exist in the mirrored repo.

What this PR does is map any relative dependencies to protocol circuits
to a git dependency before pushing that commit to the mirrored repo. It
then undoes this change before pushing to the monorepo (we want to keep
using relative paths in the monorepo).

I would've preferred using `yq` since it claims it suports TOML (and is
included in the default Github actions package list) but its support is
limited to only basic types (so no objects like `{path="..."}`) and it
can only read toml but not write it.

https://github.com/mikefarah/yq/issues/1364#issuecomment-1483957510