chore: recursion howto explainers (#3792)
# Description

This PR also moves things around to accomodate one new `how-to` and its
correspondent `explainer`

## Problem\*

An explainer and a guide on recursion was long overdue, but very much
needed.

## Summary\*

It may seem like a lot of files but most of them are just things moving
around:
- Created a new top-level category "Noir" to avoid mixing syntax stuff
with high-level explanations. That's most of the file moves
- Added a divider on the sidebar
- In the new "explanation" category, I wrote an explainer. **This is one
of the two files that needs to be reviewed**
- In the `how to` category, I wrote a guide. **This also needs to be
reviewed**
- I found out the links to `@site` were valid... But they route to the
assets, thus rendering the page in markdown format 😱 The docusaurus
documentation is not entirely clear about that, and I clearly missed the
tiny reference to that fact in the advanced pages.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>
Co-authored-by: josh crites <jc@joshcrites.com>