feat: Add context-centric based API for noir_wasm (#3798)
# Description

Currently this PR is not breaking

This adds an API that maps closer to the API we'd have in the CLI or
atleast one that differs less with the native version. One can imagine a
shared Context object and in nargo_cli the main difference is the
PathToSourceMap struct that gets passed in for example.

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>