chore: rename acir opcode from `pedersen` to `pedersen_commitment` (#3836)
# Description


Previously was a breaking change to backends, which is why it was
delayed. Since we now no longer support partial backeds, this is not a
breaking change as we are not reading an info.json file anymore, or more
precisely we are not matching on the opcodes_supported field

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.