chore: Compiler `Context` instantiates a new `CrateGraph` internally (#3847)
# Description

This is a minor cleanup task. Everywhere we created a CompilerContext,
we were passing in a default CrateGraph.

This PR just creates the default CrateGraph internally, so that the
caller does not need to. If in the future, we decide to want to pass a
populated CrateGraph, we can create a `with_crate_graph` method to
account for the edge case.

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.