chore: Rename `NpLanguage` to `ExpressionWidth` (#3834)
# Description


NpLanguage does not encapsulate what we want to describe here. We simply
want to specify an expression width and a particular backend will need
to decide as to what configuration their backend supports.

To elaborate, imagine a new np language came out which was popular;
before this PR, we may need to add in another variant. However, with
this change, the new backend will just target whatever width makes sense
for this new proving system.

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.