feat: bb uses goblin (#3636)
"vertical slice" of goblin integrating with ACIR and BB.

Beginning of goblin ultra honk interface for bb. Can only run a basic
test (assert_statement) for now and is not intended for use outside of
the CI check that it's still working, for now.
- adds bb and bb.js command `prove_and_verify_goblin`. Adds bb.js
bindings for goblin prove/verify
- modifies ACIR `dsl` folder to be able to take a goblin builder
- adds CI calls to bb.js and bb that use `prove_and_verify_goblin`
- allowing ability to load grumpkin SRS through memory, needed for bb
and (especially) bb.js. This allows an alternate source of points to be
used other than the default file-based grumpkin loader, which mostly
only works in dev (though revisit: could this work for native bb?)

---------

Co-authored-by: ledwards2225 <ledwards2225@users.noreply.github.com>
Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>
Co-authored-by: ludamad <adam@aztecprotocol.com>
Co-authored-by: ludamad <adam.domurad@gmail.com>