chore: Push `FileManager` population up to `nargo_cli` (#3844)
# Description

This tries to undo some of the changes done in #3496 where cli methods
were moved into the nargo library. This pushes the populating of the
file manager up to the cli commands, in most places this is just after
the workspace has been resolved.

A possible next step after this is to have the nargo library to no
longer ask for a file manager struct, but instead it asks for a
FileManager trait which the previous FileManager struct would implement.

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: Tom French <tom@tomfren.ch>