chore: rename `barretenberg_blackbox_solver` to `bn254_blackbox_solver` (#3874)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR renames `barretenberg_blackbox_solver` to
`bn254_blackbox_solver` and marks it as not deprecated.

This signifies that these blackbox solvers should work against all
backends targetting the bn254 curve and not just barretenberg. That
we're using barretenberg to solve them is then just an implementation
detail.

We've decided to retain the `BlackBoxFunctionSolver` trait to allow
these solvers to be swapped out based on the `FieldElement`
implementation and so the new `Bn254BlackBoxSolver` is not deprecated.


## Additional Context

Now that this solver is tied to a curve rather than a particular
backend, it's now proper to integrate it into the compiler such that we
can solve black box functions such as that which caused #3865

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.