feat: Implement DAP protocol in Nargo (#3627)
# Description

Implements the DAP protocol to allow a tool such as VS.Code to drive the
debugger in a Noir binary project. A separate PR for the [VS.Code Noir
extension](https://github.com/noir-lang/vscode-noir) will be submitted
later.

## Problem

Part of #3015 

The beginning of this implementation was heavily inspired by @dmvict
[implementation](https://github.com/noir-lang/noir/issues/3094) of an
alternative debugger for Noir programs.

## Summary

This PR implements a new `nargo` subcommand `dap`. This starts a DAP
server in single session mode through stdin/stdout and waits for a
launch request. Using the arguments in the launch request
(`projectFolder`, and optionally `package` and `proverName`) it compiles
the Noir binary package and starts it in debug mode. Through DAP
requests, a tool can then step through the program, set breakpoints and
list the generated opcodes using a disassemble request.

## Additional Context



## Documentation

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [X] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>