chore(debugger): Add guide on how to test unmerged experimental features (#3853)
# Description

There's a number of debugger features that are in active development and
that can't yet be merged to the main branch for different reasons. This
changeset adds a summary of what those features are and how to try them
out.

## Problem

Part of #3015. 

## Summary

Adds instructions on how to test experimental debugger features to the
debugger's README.md file.

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.