chore: error on cspell issues in docs (#3886)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR updates the spellcheck workflow to run a separate pass over the
docs which will error rather than raising a warning. This is as it's
more important to catch spelling issues in the docs rather than in the
code (plus the docs is going to use fewer non-standard-english library
names, etc. so false positives will be rarer.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.