fix: allow abi encoding tuples from JS (#3894)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We currently disallow abi encoding tuples which contain a mix of
primitive types, arrays and structs due to the types which have been
applied. This PR loosens the types so that we can pass a mix of these
types as a tuple.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.