fix(docs): Force docs build using latest released code always (#3762)
Having two different docs builds, one with released code and one
without, doesn't fly with build-system, since it checks the content hash
of the content to determine whether to rebuild based on committed data
only. This means that the hack introduced
[here](https://github.com/AztecProtocol/aztec-packages/pull/3716) does
not work.

So we'll just build as if we were in master on every run, for now.