feat: txpool persistence (#3672)
This PR adds a new implementation of a `TxPool` based on the kv-store
introduced in #3628. It also amends the p2p-client's bootstrap flow.

The p2p-client saves the block number its on to the database and
restores it the next time it is started. The initial sync happens as
before (it syncs to the the tip of the chain, but instead of syncing
from 0, it syncs from the last known block number) and after that, it
re-publishes any transactions in its TxPool that haven't been processed
already.

Fix #3365